[incubator-kie-issues-1522] Transition object to make extensible user task transition restendpoint
