+++ api/kogito-api/src/main/java/org/kie/kogito/usertask/model/TransitionInfo.java
@@ -1,0 +1,58 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.kie.kogito.usertask.model;
+
+import java.util.Collections;
+import java.util.Map;
+
+public class TransitionInfo {
+
+    private String transitionId;
+
+    private Map<String, Object> data;
+
+    public TransitionInfo() {
+        
+    }
+
+    public TransitionInfo(String transitionId) {
+        this(transitionId, Collections.emptyMap());
+    }
+
+    public TransitionInfo(String transitionId, Map<String, Object> data) {
+        this.transitionId = transitionId;
+        this.data = data;
+    }
+
+    public Map<String, Object> getData() {
+        return data;
+    }
+
+    public void setData(Map<String, Object> data) {
+        this.data = data;
+    }
+
+    public String getTransitionId() {
+        return transitionId;
+    }
+
+    public void setTransitionId(String transitionId) {
+        this.transitionId = transitionId;
+    }
+}
